introduced more durations and compute frequency from pitch

+++ synthesizer.hs
@@ -7,7 +7,7 @@ outputSound = B.putStr.B.pack.map fromIntegral.scale (0,255)
 main = do
   [frequency,volume,duration] <- getArgs
   let f = read frequency :: Int
-  let d = read duration :: Int
+  let d = read duration :: Double
   let a = read volume :: Double
   outputSound $ computeSound f d a
